Secured communication support for e2b.search()
Unstable with 2 or less devices but works with more devices on the bus.

+++ E2B/src/E2B.h
@@ -91,7 +91,7 @@ class E2B{
     bool hostFlag;
     bool secureFlag;
     bool isLocked;
-    uint8_t unlockedState;
+    int unlockedState;
     uint8_t secureKey;
 
     //E2Bslave
